Switch to double precision

+++ trace.h
@@ -21,14 +21,14 @@ typedef struct {
      int num_threads;
      int x;
      int y;
-     float fov;
+     double fov;
      LL* entities;
      color* output;
 } render_context;
 
 render_context* createRenderContext();
 void addEntity(render_context* rc, entity* e);
-entity* createSphere(float4 pos, float r, material m);
+entity* createSphere(float4 pos, double r, material m);
 entity* createPlane(float4 pos, float4 norm, material m1, material m2);
 void renderScene(render_context* c, int x, int y, int num_threads);
